refactor(ui/components): route imports through barrels
