+++ src/components/ConfirmDialog.tsx
@@ -1,7 +1,6 @@
 import { X } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
-import { Button } from '@/components/ui/Button'
-import { Modal } from '@/components/ui/Modal'
+import { Button, Modal } from '@/components/ui'
 
 interface ConfirmDialogProps {
     title: string
