+++ src/components/screens/index.ts
@@ -1,0 +1,4 @@
+export { AppsScreen } from './AppsScreen'
+export { HistoryScreen } from './HistoryScreen'
+export { LogScreen } from './LogScreen'
+export { SettingsScreen } from './SettingsScreen'
